Added Finalized-parameter

+++ Backing/EncounterForm.razor.cs
@@ -30,7 +30,7 @@ namespace RaidPlannerClient.Components
 
         [Inject]
         public IEncounterService EncounterService { get; set; }
-        
+
         [Inject]
         private IJSRuntime JSRuntime { get; set; }
 
@@ -44,21 +44,27 @@ namespace RaidPlannerClient.Components
         public Character GetCharacterById(int? id)
         {
             Console.WriteLine($"Getting character with id {id}");
-            if(Players.SelectMany(p=>p.Characters).Any(c=>c.Id==id)) {
+            if (Players.SelectMany(p => p.Characters).Any(c => c.Id == id))
+            {
                 return Players.SelectMany(p => p.Characters).Where(c => c.Id == id).FirstOrDefault();
-            } else {
-                return new Character{Id=id, Name="Unknown"};
+            }
+            else
+            {
+                return new Character { Id = id, Name = "Unknown" };
             };
         }
 
         public Boss GetBossById(int? id)
         {
             Console.WriteLine($"Getting boss with id {id}");
-            if(Instances.SelectMany(i=>i.Bosses).Any(b=>b.Id==id)) {
+            if (Instances.SelectMany(i => i.Bosses).Any(b => b.Id == id))
+            {
                 return Instances.SelectMany(i => i.Bosses).Where(b => b.Id == id).FirstOrDefault();
-            } else {
+            }
+            else
+            {
                 Console.WriteLine($"Boss with id {id} does not exist, returning unknown boss");
-                return new Boss{Id=id, Name="Unknown"};
+                return new Boss { Id = id, Name = "Unknown" };
             }
         }
 
@@ -96,35 +102,45 @@ namespace RaidPlannerClient.Components
 
         public void Add(Player player, Character character, Role role)
         {
-            Console.WriteLine($"Adding {player.Name} as {role}");
-            var encounterCharacter = new EncounterCharacter { PlayerId = (int)player.Id, CharacterId = (int)character.Id, Role = role };
-            EncounterService.AddCharacter(Raid, Encounter, encounterCharacter);
-            Encounter.Characters.Add(encounterCharacter);
+            if (!Raid.Finalized)
+            {
+                Console.WriteLine($"Adding {player.Name} as {role}");
+                var encounterCharacter = new EncounterCharacter { PlayerId = (int)player.Id, CharacterId = (int)character.Id, Role = role };
+                EncounterService.AddCharacter(Raid, Encounter, encounterCharacter);
+                Encounter.Characters.Add(encounterCharacter);
+            }
         }
 
         public void Remove(int playerId)
         {
-            var player = Players.Where(p => p.Id == playerId).First();
-            Console.WriteLine($"Removing {player.Name}");
-            EncounterService.DeleteCharacter(Raid, Encounter, Encounter.Characters.First(p => p.PlayerId == playerId));
-            Encounter.Characters.RemoveAll(p => p.PlayerId == playerId);
+            if (!Raid.Finalized)
+            {
+                var player = Players.Where(p => p.Id == playerId).First();
+                Console.WriteLine($"Removing {player.Name}");
+                EncounterService.DeleteCharacter(Raid, Encounter, Encounter.Characters.First(p => p.PlayerId == playerId));
+                Encounter.Characters.RemoveAll(p => p.PlayerId == playerId);
+            }
         }
 
         public List<EncounterCharacter> GetPlayersByRole(Role role)
         {
             return Encounter.Characters
                     .Where(p => p.Role == role)
-                    .Where(p=>Players.SelectMany(p=>p.Characters).Any(c=>c.Id==p.CharacterId))
+                    .Where(p => Players.SelectMany(p => p.Characters).Any(c => c.Id == p.CharacterId))
                     .OrderBy(p => GetCharacterById(p.CharacterId).Name)
                     .ToList();
         }
 
         public async void DeleteEncounter()
         {
-            bool confirmed = await JSRuntime.InvokeAsync<bool>("confirm", $"Are you sure you want to delete encounter {GetBossById(Encounter.BossId).Name}?");
-            if(confirmed) {
-                await EncounterService.DeleteEncounter(Raid, Encounter);
-                RaidForm.DeleteEncounter(Encounter);
+            if (!Raid.Finalized)
+            {
+                bool confirmed = await JSRuntime.InvokeAsync<bool>("confirm", $"Are you sure you want to delete encounter {GetBossById(Encounter.BossId).Name}?");
+                if (confirmed)
+                {
+                    await EncounterService.DeleteEncounter(Raid, Encounter);
+                    RaidForm.DeleteEncounter(Encounter);
+                }
             }
         }
     }
